feat: drop 확인창 개발 중
