+++ packages/ui/panda/package.json
@@ -11,7 +11,9 @@
   "dependencies": {
     "@gitanimals/react": "workspace:*",
     "@radix-ui/react-accordion": "^1.2.1",
+    "@radix-ui/react-checkbox": "^1.1.1",
     "@radix-ui/react-dialog": "^1.1.2",
+    "@radix-ui/react-label": "^2.1.1",
     "@radix-ui/react-select": "^2.1.2",
     "@shadow-panda/style-context": "^0.7.1",
     "lucide-react": "^0.408.0"
