+++ packages/ui/panda/src/components/CheckBox/index.ts
@@ -1,0 +1,1 @@
+export { Checkbox } from './CheckBox';
