+++ packages/ui/panda/src/components/Label/Label.tsx
@@ -1,0 +1,5 @@
+import { styled, type HTMLStyledProps } from '_panda/jsx';
+import { label } from '_panda/recipes';
+
+export const Label = styled('label', label);
+export type LabelProps = HTMLStyledProps<typeof Label>;
