+++ packages/ui/panda/src/components/Label/index.tsx
@@ -1,0 +1,1 @@
+export { Label } from './Label';
